fix(dust-hive): remove code duplication, dead code, and update docs
- Extract isErrnoException to shared lib/errors.ts (was duplicated 6x)
- Extract loadEnvVars to shared lib/env-utils.ts (was duplicated 2x)
- Export isServiceName from services.ts and shellQuote from shell.ts
- Remove duplicate SEED_USER_PATH from seed-config.ts
- Delete unused diff.ts module
- Remove unused exports: initializeCache, runCachedBinary, promptYesNo,
  promptChoice, usageError, getServicePort
- Remove unused imports from destroy.ts
- Fix non-null assertion in seed.ts with for-of loop
- Update README.md and CLAUDE.md to match actual CLI implementation
