+++ x/henry/dust-hive/CLAUDE.md
@@ -48,7 +48,7 @@ src/
 ├── index.ts           # CLI entry point
 ├── forward-daemon.ts  # TCP forwarder daemon (ports 3000,3001,3002,3006 → env)
 ├── commands/          # Command implementations (all MVP commands complete)
-│   ├── cache.ts       # Cache management (show/rebuild binaries)
+│   ├── cache.ts       # Cache management (show cache status)
 │   ├── cool.ts        # Stop services, keep SDK
 │   ├── destroy.ts     # Remove environment
 │   ├── doctor.ts      # Prerequisite checking
@@ -58,19 +58,24 @@ src/
 │   ├── open.ts        # Attach to zellij session
 │   ├── reload.ts      # Kill and reopen zellij session
 │   ├── restart.ts     # Restart a single service
+│   ├── seed-config.ts # Extract user data from existing DB
 │   ├── spawn.ts       # Create environment
 │   ├── start.ts       # Resume stopped env
 │   ├── status.ts      # Show service health
 │   ├── stop.ts        # Full stop
+│   ├── sync.ts        # Pull main, rebuild binaries, refresh deps
 │   ├── url.ts         # Print front URL
 │   └── warm.ts        # Start docker + all services
 └── lib/               # Shared utilities
+    ├── activity.ts    # Last-active environment tracking
     ├── cache.ts       # Binary caching for fast init
     ├── commands.ts    # Command helpers (requireEnvironment)
     ├── config.ts      # Configuration management
     ├── docker.ts      # Docker compose + start/stop operations
     ├── environment.ts # Environment CRUD
+    ├── env-utils.ts   # Environment variable loading
     ├── envgen.ts      # env.sh generation
+    ├── errors.ts      # Shared error utilities
     ├── forwarderConfig.ts # Forwarder port mappings
     ├── fs.ts          # Filesystem helpers
     ├── forward.ts     # TCP forwarder management
@@ -79,8 +84,10 @@ src/
     ├── paths.ts       # Path constants and helpers
     ├── ports.ts       # Port allocation
     ├── process.ts     # Daemon management (PID files, spawn/kill)
+    ├── prompt.ts      # Interactive prompts
     ├── registry.ts    # Service registry (config, health checks)
     ├── result.ts      # Result<T,E> type for error handling
+    ├── seed.ts        # SQL-based database seeding
     ├── services.ts    # Service names and types
     ├── setup.ts       # Dependency installation
     ├── shell.ts       # Shell command builder
@@ -89,7 +96,7 @@ src/
     └── worktree.ts    # Git worktree operations
 
 tests/
-└── lib/               # Unit tests for lib modules (156 tests)
+└── lib/               # Unit tests for lib modules
     ├── docker.test.ts
     ├── environment.test.ts
     ├── envgen.test.ts
@@ -122,7 +129,7 @@ tests/
 
 | Command | Description |
 |---------|-------------|
-| `spawn` | Create environment (worktree + symlinks + SDK watch); supports --warm to open a warm tab, --no-attach to run in background |
+| `spawn` | Create environment (worktree + symlinks + SDK watch); supports --warm, --no-attach, --wait |
 | `warm` | Start docker + all services (auto-forwards port 3000, supports --no-forward/--force-ports) |
 | `cool` | Stop services, keep SDK watch |
 | `start` | Resume stopped env |
@@ -136,9 +143,10 @@ tests/
 | `logs` | View service logs |
 | `url` | Print front URL |
 | `doctor` | Check prerequisites |
-| `cache` | Show or rebuild binary cache |
+| `cache` | Show binary cache status |
 | `forward` | Manage OAuth port forwarding (ports 3000,3001,3002,3006 → env) |
-| `sync [BRANCH]` | Rebase on branch (default: main), rebuild binaries, refresh node_modules |
+| `sync` | Pull latest main, rebuild binaries, refresh deps |
+| `seed-config` | Extract user data from existing DB for seeding new environments |
 
 ## Performance
 
@@ -172,9 +180,7 @@ caches compilations by content hash, making rebuilds after branch switches faste
 Check cache status:
 ```bash
 dust-hive cache            # Show what's cached
-dust-hive cache status     # Explicit status
-dust-hive cache --rebuild  # Build missing binaries
-dust-hive cache rebuild    # Alias for rebuild
+dust-hive sync             # Pull main, rebuild binaries, refresh deps
 ```
 
 ## Testing the CLI
