+++ x/henry/dust-hive/src/lib/cache.ts
@@ -137,77 +137,3 @@ export async function buildBinaries(
 
   return { success: failed.length === 0, built, failed };
 }
-
-// Initialize cache from a repo (auto-set source, optionally build missing)
-export async function initializeCache(
-  repoRoot: string,
-  options?: { rebuild?: boolean }
-): Promise<{
-  source: string;
-  available: Binary[];
-  missing: Binary[];
-}> {
-  await setCacheSource(repoRoot);
-
-  if (options?.rebuild) {
-    // Rebuild all binaries
-    logger.step("Rebuilding all binaries...");
-    const result = await buildBinaries(repoRoot, [...ALL_BINARIES]);
-    return {
-      source: repoRoot,
-      available: result.built,
-      missing: result.failed,
-    };
-  }
-
-  const available = await getAvailableBinaries(repoRoot);
-  const missing = await getMissingBinaries(repoRoot);
-
-  return { source: repoRoot, available, missing };
-}
-
-// Run a cached binary (with fallback to cargo run)
-export async function runCachedBinary(
-  cacheSource: string,
-  binary: Binary,
-  args: string[],
-  options: {
-    cwd: string;
-    env?: Record<string, string>;
-    fallbackToCargo?: boolean;
-  }
-): Promise<{ success: boolean; usedCache: boolean }> {
-  const binaryPath = getBinaryPath(cacheSource, binary);
-  const hasBinary = await binaryExists(cacheSource, binary);
-
-  if (hasBinary) {
-    // Run the cached binary directly
-    const proc = Bun.spawn([binaryPath, ...args], {
-      cwd: options.cwd,
-      env: { ...process.env, ...options.env },
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-
-    await proc.exited;
-    return { success: proc.exitCode === 0, usedCache: true };
-  }
-
-  if (options.fallbackToCargo !== false) {
-    // Fall back to cargo run
-    const cargoArgs = ["cargo", "run", "--bin", binary, "--", ...args].join(" ");
-    const command = buildShell({ run: cargoArgs });
-
-    const proc = Bun.spawn(["bash", "-c", command], {
-      cwd: options.cwd,
-      env: { ...process.env, ...options.env },
-      stdout: "pipe",
-      stderr: "pipe",
-    });
-
-    await proc.exited;
-    return { success: proc.exitCode === 0, usedCache: false };
-  }
-
-  return { success: false, usedCache: false };
-}
