+++ x/henry/dust-hive/src/lib/ports.ts
@@ -2,6 +2,7 @@ import { spawnSync } from "node:child_process";
 import { mkdir, open, readdir, stat, unlink } from "node:fs/promises";
 import { join } from "node:path";
 import { z } from "zod";
+import { isErrnoException } from "./errors";
 import { directoryExists } from "./fs";
 import { DUST_HIVE_ENVS, DUST_HIVE_HOME, getPortsPath } from "./paths";
 import { isProcessRunning, killProcess } from "./process";
@@ -46,10 +47,6 @@ const PortAllocationFields = z.object({
 const PortAllocationSchema = PortAllocationFields.passthrough();
 
 export type PortAllocation = z.infer<typeof PortAllocationFields>;
-
-function isErrnoException(error: unknown): error is NodeJS.ErrnoException {
-  return typeof error === "object" && error !== null && "code" in error;
-}
 
 // Type guard for PortAllocation
 function isPortAllocation(data: unknown): data is PortAllocation {
