+++ x/henry/dust-hive/src/lib/seed.ts
@@ -7,6 +7,7 @@
 // The SQL is defined in front/lib/dev/dust_hive_seed.sql
 // This file only handles sId generation and parameter passing.
 
+import { loadEnvVars } from "./env-utils";
 import type { Environment } from "./environment";
 import { logger } from "./logger";
 import { SEED_USER_PATH, getEnvFilePath, getWorktreeDir } from "./paths";
@@ -17,9 +18,8 @@ function generateRandomModelSId(): string {
   const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
   const bytes = crypto.getRandomValues(new Uint8Array(10));
   let result = "";
-  for (let i = 0; i < 10; i++) {
-    // biome-ignore lint/style/noNonNullAssertion: bytes[i] is always defined for i < 10
-    result += chars[bytes[i]! % chars.length];
+  for (const byte of bytes) {
+    result += chars[byte % chars.length];
   }
   return result;
 }
@@ -44,32 +44,6 @@ async function loadSeedConfig(): Promise<SeedConfig | null> {
     return null;
   }
   return file.json();
-}
-
-// Load environment variables from env.sh
-async function loadEnvVars(envShPath: string): Promise<Record<string, string>> {
-  const command = `source "${envShPath}" && env`;
-  const proc = Bun.spawn(["bash", "-c", command], {
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-
-  const output = await new Response(proc.stdout).text();
-  const stderr = await new Response(proc.stderr).text();
-  await proc.exited;
-
-  if (proc.exitCode !== 0) {
-    throw new Error(`Failed to load env vars: ${stderr.trim() || "unknown error"}`);
-  }
-
-  const env: Record<string, string> = {};
-  for (const line of output.split("\n")) {
-    const idx = line.indexOf("=");
-    if (idx > 0) {
-      env[line.substring(0, idx)] = line.substring(idx + 1);
-    }
-  }
-  return env;
 }
 
 function buildDatabaseUri(envVars: Record<string, string>): string {
