+++ x/henry/dust-hive/src/lib/services.ts
@@ -9,3 +9,10 @@ export const ALL_SERVICES = [
 ] as const;
 
 export type ServiceName = (typeof ALL_SERVICES)[number];
+
+/**
+ * Type guard to check if a string is a valid service name.
+ */
+export function isServiceName(value: string | undefined): value is ServiceName {
+  return value !== undefined && ALL_SERVICES.includes(value as ServiceName);
+}
